refactor: Rename template to builders
